+++ src/tooling/piral-cli/src/apps/run-emulator-piral.ts
@@ -117,7 +117,7 @@ always-auth=true`,
 
   const npmClient = await determineNpmClient(appRoot, defaultNpmClient);
   const [packageName] = await installPiralInstance(app, fullBase, appRoot, npmClient);
-  const piral = findPiralInstance(packageName, appRoot, originalPort);
+  const piral = await findPiralInstance(packageName, appRoot, originalPort);
   const port = await getAvailablePort(piral.port);
 
   const krasBaseConfig = resolve(fullBase, krasrc);