+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Piral Changelog
 
+## 0.15.4 (tbd)
+
+- Fixed issue loading `dotnetjs` in `piral-blazor`
+- Added watcher for *package.json* and config files to `piral debug`
+- Added watcher for *package.json* and config files to `pilet debug`
+- Added `--watch` flag to `piral build` command to `piral-cli`
+- Added `--watch` flag to `pilet build` command to `piral-cli`
+
 ## 0.15.3 (December 16, 2022)
 
 - Added support for Blazor custom elements in `piral-blazor`